+++ packages/css/src/components/tile/tile.stories.ts
@@ -1,0 +1,16 @@
+import { storiesOfTile } from '@dso-toolkit/sources';
+import { storiesOf } from '@storybook/web-components'
+
+import { tileTemplate } from './tile.template';
+import readme from './readme.md';
+
+storiesOfTile(
+  {
+    module,
+    storiesOf,
+    readme
+  },
+  {
+    tileTemplate
+  }
+);