+++ packages/css/src/components/whitebox/whitebox.stories.ts
@@ -1,0 +1,16 @@
+import { storiesOfWhitebox } from '@dso-toolkit/sources';
+import { storiesOf } from '@storybook/web-components'
+
+import { whiteboxTemplate } from './whitebox.template';
+import readme from './readme.md';
+
+storiesOfWhitebox(
+  {
+    module,
+    storiesOf,
+    readme
+  },
+  {
+    whiteboxTemplate
+  }
+);