+++ packages/sources/src/stories-helpers/index.ts
@@ -67,3 +67,7 @@ export type ArgTypes<ComponentArgs> = {
 
   return [checked];
 }
+
+export function componentArgs<T>(args: T): T {
+  return args;
+}